requirements.txt
streamlit>=1.28
python-dotenv>=0.19
google-generativeai>=0.3
tensorflow>=2.10
tensorflow-tts>=0.6
vosk>=0.3.45
pydub>=0.25
soundfile>=0.12
numpy>=1.22